Calypso keypad driver
This driver provides a char device to access the keypad matrix HW built into
the Calypso DBB. Pressing/releasing of keys is send as scancodes. A to Y are
assigned to the 25 matrix buttons, Z represents 'Power/End call'. When pushed
down, the lower-case letter is send. On return to the 'upper' position, the
respective upper-case letter is send. With the exception of the power button,
any set may be pressed simultanously.